split up files, sort entries, change formatting
